generate_custom_video/Stable-Disffsion/requirements.txt
fastapi
pydantic
python-dotenv
oss2
httpx
uvicorn
ffmpeg-python